Modified the MM plugin and fastboot plugin to be with fm101
In fu-mm-device.c for fm101, Added probe_fibocom.
T version was by the AT command, and six child device were
created to record version and mccmnc.

In fu-mm-plugin.c, detach checks if the module is fibocom,
uninhibit_device is executed when the module is obtained again.
Because other modem policies are not certain, This is only for
fibocom.

In fu-fastboot-device.c, write_firmware detects that the firmware
contains a file named flash_end, which is used to mark the
completion of the modem flash. This is only for fibocom.